修正 CSV 課程匯入時節數欄位會因為 parseInt 而將 0.5 節轉為 1 節的 Bug

+++ static/app.js
@@ -1419,7 +1419,7 @@ async function processCSVImport(text, type) {
             entries.push({
                 class_name: className,
                 subject: subject,
-                periods: parseInt(cols[2]) || 1,
+                periods: parseFloat(cols[2]) || 1,
                 teacher_name: (cols[3] || "").trim(),
                 classroom_name: (cols[4] || "").trim() || null,
                 week_type: (cols[5] || "EVERY").trim().toUpperCase()
